Recover panics in finishRequest, write correct API response
